backend/api/requirements.txt
anyio==4.8.0
astroid==3.3.9
attrs==25.1.0
azure-core==1.32.0
azure-cosmos==4.9.0
azure-functions==1.21.3
azure-identity==1.19.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
dill==0.3.9
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.1
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
outcome==1.3.0.post0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.5
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-dotenv==0.5.2
python-dotenv==1.0.1
pywin32==308
requests==2.32.3
rsa==4.9
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.2
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
